Rename package from api to query (#103)
